Move ImproveMemoryOps to separate file
