+++ src/opt/ImproveMemoryOps.h
@@ -1,0 +1,63 @@
+/*
+  Copyright (c) 2022, Intel Corporation
+  All rights reserved.
+
+  Redistribution and use in source and binary forms, with or without
+  modification, are permitted provided that the following conditions are
+  met:
+
+    * Redistributions of source code must retain the above copyright
+      notice, this list of conditions and the following disclaimer.
+
+    * Redistributions in binary form must reproduce the above copyright
+      notice, this list of conditions and the following disclaimer in the
+      documentation and/or other materials provided with the distribution.
+
+    * Neither the name of Intel Corporation nor the names of its
+      contributors may be used to endorse or promote products derived from
+      this software without specific prior written permission.
+
+
+   THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
+   IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
+   TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+   PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER
+   OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+   EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+   PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+   PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+   LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+   NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+   SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+*/
+
+#pragma once
+
+#include "ISPCPass.h"
+
+namespace ispc {
+
+/** When the front-end emits gathers and scatters, it generates an array of
+    vector-width pointers to represent the set of addresses to read from or
+    write to.  This optimization detects cases when the base pointer is a
+    uniform pointer or when the indexing is into an array that can be
+    converted into scatters/gathers from a single base pointer and an array
+    of offsets.
+
+    See for example the comments discussing the __pseudo_gather functions
+    in builtins.cpp for more information about this.
+ */
+class ImproveMemoryOpsPass : public llvm::FunctionPass {
+  public:
+    static char ID;
+    ImproveMemoryOpsPass() : FunctionPass(ID) {}
+
+    llvm::StringRef getPassName() const { return "Improve Memory Ops"; }
+
+    bool runOnBasicBlock(llvm::BasicBlock &BB);
+
+    bool runOnFunction(llvm::Function &F);
+};
+
+llvm::Pass *CreateImproveMemoryOpsPass();
+} // namespace ispc
